refactor: modularize project running
Move running logic to its own module file for better organization,
splitting of functions and more unit-testable code.
